requirements.txt
asgiref==3.5.0
cssselect2==0.6.0
Django==3.1.14
django-classy-tags==3.0.1
django-cms==3.8.0
django-filer==2.1.2
django-formtools==2.3
django-js-asset==2.0.0
django-markdownx==3.0.1
django-mptt==0.13.4
django-polymorphic==3.0.0
django-sekizai==3.0.1
django-tables2==2.4.1
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.1.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.8.1
html5lib==1.1
importlib-metadata==4.11.3
lxml==4.8.0
Markdown==3.3.6
Pillow==9.1.0
pytz==2022.1
reportlab==3.6.9
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
Unidecode==1.1.2
webencodings==0.5.1
zipp==3.8.0
